Desarrollo, mejora seguridad dentro de ficheros, agregado el panel de admin para administrar los admins de la matriz

+++ Admin/src/class/classDatabase.php
@@ -2,15 +2,15 @@
 
 class Database{
 	//LOCAL
-	/*private $dbHost 	= "localhost";
+	private $dbHost 	= "localhost";
 	private $dbUser 	= "root";
 	private $dbPassword = "root";
-	private $dbDatabase = "matriz";*/
+	private $dbDatabase = "matriz";
 	
-	private $dbHost 	= "localhost";
+	/*private $dbHost 	= "localhost";
 	private $dbUser 	= "matrizroot";
 	private $dbPassword = "Matriz159!!";
-	private $dbDatabase = "matriz";
+	private $dbDatabase = "matriz";*/
 	
 	private $dbLink      = "";
 	private $dbRecordSet = 0;
@@ -34,8 +34,8 @@ class Database{
 	
 	//Conexion
 	public function conect(){
-		//$this->dbLink = mysql_connect($this->dbHost, $this->dbUser) or die ("1. No funciona por " . mysql_error()); 
-		$this->dbLink= mysql_connect($this->dbHost, $this->dbUser, $this->dbPassword) or die ("1. No funciona por " . mysql_error()); 
+		$this->dbLink = mysql_connect($this->dbHost, $this->dbUser) or die ("1. No funciona por " . mysql_error()); 
+		//$this->dbLink= mysql_connect($this->dbHost, $this->dbUser, $this->dbPassword) or die ("1. No funciona por " . mysql_error()); 
 	}
 	//Seleccionar base	
 	public function setBase(){
@@ -49,8 +49,8 @@ class Database{
 	//Desconexion
 	private function disconnect(){
 		if($this->dbLink){
-			//mysql_close( mysql_connect($this->dbHost, $this->dbUser) );
-			mysql_close( mysql_connect($this->dbHost, $this->dbUser, $this->dbPassword) );
+			mysql_close( mysql_connect($this->dbHost, $this->dbUser) );
+			//mysql_close( mysql_connect($this->dbHost, $this->dbUser, $this->dbPassword) );
 			//mysql_close($this->dbLink);
 		}
 	}
@@ -193,7 +193,7 @@ class Database{
 	public function DeleteImagen($link){
 		
 		//no elimina las imagenes por defecto
-		if($link != 'images/es.png' && $link != '../images/es.png'){
+		if($link != 'images/es.png' && $link != '../images/es.png' && file_exists($link) ){
 			if(unlink($link)){
 				return true;
 			}else{
